qartodRunner class, qartod module, pulls in basic qartod tables
