+++ rca_data_tools/qaqc/dashboard.py
@@ -36,6 +36,7 @@ import xml.etree.ElementTree as et
 
 from rca_data_tools.qaqc.utils import select_logger, save_fig, get_s3_kwargs
 from rca_data_tools.qaqc.constants import variable_paramDict, statusColors, discreteSample_dict
+from rca_data_tools.qaqc.calculate import QartodRunner
 INPUT_BUCKET = "ooi-data/"
 
 
@@ -279,108 +280,6 @@ def loadProfiles(refDes):
     return profileList
 
 
-
-
-def loadQARTOD(refDes, param, sensorType, logger=select_logger()):
-
-    renameMap = {
-                 'sea_water_temperature':'seawater_temperature',
-                 'sea_water_practical_salinity':'practical_salinity',
-                 'sea_water_pressure':'seawater_pressure',
-                 'sea_water_density':'density',
-                 #'ph_seawater':'seawater_ph',
-                 }
-
-    if param in renameMap:
-        param = renameMap[param]
-
-    (site, node, sensor1, sensor2) = refDes.split('-')
-    sensor = sensor1 + '-' + sensor2
-
-    # if parameter is oxygen sensor in ctd stream, replace sensor type
-    if sensorType == 'ctdpf' and 'oxygen' in param:
-        if 'SF' in node:
-            sensorType = 'dofst'
-        else:
-            sensorType = 'dosta'
-
-    # Load climatology and gross range values
-
-    githubBaseURL = 'https://raw.githubusercontent.com/oceanobservatories/qc-lookup/master/qartod/'
-
-    if sensorType == 'phsen':
-        param = 'seawater_ph'
-
-    clim_URL = (
-        githubBaseURL
-        + sensorType
-        + '/climatology_tables/'
-        + refDes
-        + '-'
-        + param
-        + '.csv'
-    )
-    grossRange_URL = (
-        githubBaseURL
-        + sensorType
-        + '/'
-        + sensorType
-        + '_qartod_gross_range_test_values.csv'
-    )
-
-    download = requests.get(grossRange_URL)
-    if download.status_code == 200:
-        df_grossRange = pd.read_csv(
-            io.StringIO(download.content.decode('utf-8'))
-        )
-        qcConfig = df_grossRange.qcConfig[
-            (df_grossRange.subsite == site)
-            & (df_grossRange.node == node)
-            & (df_grossRange.sensor == sensor)
-            & (df_grossRange.parameters.str.contains(param))
-        ]
-        if len(qcConfig) > 0:
-            qcConfig_json = qcConfig.values[0].replace("'", "\"")
-            grossRange_dict = json.loads(qcConfig_json)
-        else:
-            logger.warning(
-                f"error retrieving gross range data for {refDes} {param} {sensorType}"
-            )
-            grossRange_dict = {}
-    else:
-        logger.warning(
-            f"error retrieving gross range data for {refDes} {param} {sensorType}"
-        )
-        grossRange_dict = {}
-
-    download = requests.get(clim_URL)
-    if download.status_code == 200:
-        df_clim = pd.read_csv(io.StringIO(download.content.decode('utf-8')))
-        climRename = {
-            'Unnamed: 0': 'depth',
-            '[1, 1]': '1',
-            '[2, 2]': '2',
-            '[3, 3]': '3',
-            '[4, 4]': '4',
-            '[5, 5]': '5',
-            '[6, 6]': '6',
-            '[7, 7]': '7',
-            '[8, 8]': '8',
-            '[9, 9]': '9',
-            '[10, 10]': '10',
-            '[11, 11]': '11',
-            '[12, 12]': '12',
-        }
-
-        df_clim.rename(columns=climRename, inplace=True)
-        clim_dict = df_clim.set_index('depth').to_dict()
-    else:
-        logger.warning(
-            f"error retrieving climatology data for {refDes} {param} {sensorType}"
-        )
-        clim_dict = {}
-
-    return (grossRange_dict, clim_dict)
 
 def loadStatus():
     statusResponse = requests.get("https://nereus.ooirsn.uw.edu/api/public/v1/instruments/operational-status").text
@@ -1309,13 +1208,17 @@ def plotProfilesScatter(
         elif 'flag' in overlay:
             qcDS = overlayData_flag.sel(time=slice(timeSpan[0], timeSpan[1]))
             qcDS = retrieve_qc(qcDS)
-             # TODO if homebrew_qartod: block goes here, overwriting qcDS with homebrew qartod array
             flags = {
                     'qartod_grossRange':{'symbol':'+', 'param':'_qartod_executed_gross_range_test'},
                     'qartod_climatology':{'symbol':'x','param':'_qartod_executed_climatology_test'},
                     #'qartod_summary':{'symbol':'1','param':'_qartod_results'},
                     'qc':{'symbol':'s','param':'_qc_summary_flag'},
                 }
+            
+            # TODO if homebrew_qartod: block goes here, overwriting qcDS with homebrew qartod array
+            # if True: 
+            #     qartodRunner = QartodRunner(site, param, paramData, qcDS, flags)
+
             for flagType in flags.keys():
                 flagString = Xparam + flags[flagType]['param']
                 if flagString in qcDS:
@@ -1769,7 +1672,7 @@ def plotProfilesScatter(
 @task
 def plotScatter(
     Yparam,
-    param,
+    param, # parameter short name
     paramData,
     plotTitle,
     yLabel,
@@ -2225,13 +2128,17 @@ def plotScatter(
                 print(qcDS)
                 # retrieve flags
                 qcDS = retrieve_qc(qcDS)
-                # TODO if homebrew_qartod: block goes here, overwriting qcDS with homebrew qartod array
                 flags = {
                     'qartod_grossRange':{'symbol':'+', 'param':'_qartod_executed_gross_range_test'},
                     'qartod_climatology':{'symbol':'x','param':'_qartod_executed_climatology_test'},
                     #'qartod_summary':{'symbol':'1','param':'_qartod_results'},
                     'qc':{'symbol':'s','param':'_qc_summary_flag'},
                 }
+
+                # TODO if homebrew_qartod: block goes here, overwriting qcDS with homebrew qartod array
+                if True: 
+                    qartodRunner = QartodRunner(site, Yparam, paramData, qcDS, flags)
+
                 for flagType in flags.keys():
                     flagString = Yparam + flags[flagType]['param']
                     print(flagString)
